requirements.txt
asgiref==3.8.1
Django==5.1.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.48
psycopg==3.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
